Always mock the fetch API.
